requirement.txt
Flask==2.0.1
pandas==1.3.3
openpyxl==3.0.7
biopython==1.79
joblib==1.1.0
Python==3.9.13